+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added bunch of new labels
 - Renamed `buildCaseStatement.sh` to `buildLabel.sh` and improved it a lot. It is a great start when figuring out how to create a new label for an app, or a piece of software.
 - Mosyle changed their app name from Business to Self-Service
+- Added variable `TIMEOUT` that can be set to the number of seconds after which the dialogue box will give up waiting for a button click. This prevents the script from blocking other processes while the user ignores the dialogue box for a prolonged period of time.
 
 ## v0.6 - 2021-07-14
 
@@ -33,7 +34,6 @@
 - Script `buildCaseStatement.sh` to help with creating labels have been improved.
 - Fixed a bug in a variable name that prevented updateTool to be used
 - added `type` variable for value `"updateronly"` if the label should only run an updater tool.
-- Added variable `TIMEOUT` that can be set to the number of seconds after which the dialogue box will give up waiting for a button click. This prevents the script from blocking other processes while the user ignores the dialogue box for a prolonged period of time.
 
 
 ## v0.4 - 2020-10-19